+++ metrics-annotation/src/main/java/com/yammer/metrics/annotation/ExceptionMetered.java
@@ -42,8 +42,22 @@ import java.util.concurrent.TimeUnit;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface ExceptionMetered {
-
+    /**
+     * The default suffix for meter names.
+     */
     String DEFAULT_NAME_SUFFIX = "Exceptions";
+
+    /**
+     * The group of the timer. If not specified, the meter will be given a group based on
+     * the package.
+     */
+    String group() default "";
+
+    /**
+     * The type of the timer. If not specified the meter will be given a type based on
+     * the class name.
+     */
+    String type() default "";
 
     /**
      * The name of the meter. If not specified, the meter will be given a name based on the method