+++ metrics-guice/src/test/java/com/yammer/metrics/guice/tests/GaugeTest.java
@@ -1,5 +1,7 @@
 package com.yammer.metrics.guice.tests;
 
+import java.util.Set;
+
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.yammer.metrics.core.Gauge;
@@ -8,6 +10,8 @@ import com.yammer.metrics.core.MetricName;
 import com.yammer.metrics.core.MetricsRegistry;
 import com.yammer.metrics.guice.InstrumentationModule;
 
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 import static org.hamcrest.Matchers.instanceOf;
@@ -16,14 +20,30 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertThat;
 
 public class GaugeTest {
+    private InstrumentedWithGauge instance;
+    private MetricsRegistry registry;
+
+    @Before
+    public void setup() {
+        this.registry = new MetricsRegistry();
+        final Injector injector = Guice.createInjector(new InstrumentationModule() {
+            @Override
+            protected MetricsRegistry createMetricsRegistry() {
+                return registry;
+            }
+        });
+        this.instance = injector.getInstance(InstrumentedWithGauge.class);
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        registry.shutdown();
+    }
+
 
     @Test
     @SuppressWarnings("unchecked")
     public void aGaugeAnnotatedMethod() throws Exception {
-        final Injector injector = Guice.createInjector(new InstrumentationModule());
-        final InstrumentedWithGauge instance = injector.getInstance(InstrumentedWithGauge.class);
-        final MetricsRegistry registry = injector.getInstance(MetricsRegistry.class);
-
         instance.doAThing();
 
         final Metric metric = registry.allMetrics().get(new MetricName(InstrumentedWithGauge.class,
@@ -46,10 +66,6 @@ public class GaugeTest {
     @Test
     @SuppressWarnings("unchecked")
     public void aGaugeAnnotatedMethodWithDefaultName() throws Exception {
-        final Injector injector = Guice.createInjector(new InstrumentationModule());
-        final InstrumentedWithGauge instance = injector.getInstance(InstrumentedWithGauge.class);
-        final MetricsRegistry registry = injector.getInstance(MetricsRegistry.class);
-
         instance.doAnotherThing();
 
         final Metric metric = registry.allMetrics().get(new MetricName(InstrumentedWithGauge.class,
@@ -66,5 +82,30 @@ public class GaugeTest {
         assertThat("Guice creates a gauge with the given value",
                    ((Gauge<String>) metric).value(),
                    is("anotherThing"));
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    public void aGaugeAnnotatedMethodWithGroupTypeAndName() throws Exception {
+        final Injector injector = Guice.createInjector(new InstrumentationModule());
+        final InstrumentedWithGauge instance = injector.getInstance(InstrumentedWithGauge.class);
+        final MetricsRegistry registry = injector.getInstance(MetricsRegistry.class);
+
+        instance.doAThingWithGroupTypeAndName();
+
+        Set<MetricName> keySet = registry.allMetrics().keySet();
+        final Metric metric = registry.allMetrics().get(new MetricName("g", "t", "n"));
+
+        assertThat("Guice creates a metric",
+                   metric,
+                   is(notNullValue()));
+
+        assertThat("Guice creates a gauge",
+                   metric,
+                   is(instanceOf(Gauge.class)));
+
+        assertThat("Guice creates a gauge with the given value",
+                   ((Gauge<String>) metric).value(),
+                   is("anotherThingWithGroupTypeAndName"));
     }
 }