+++ metrics-spring/src/main/java/com/yammer/metrics/spring/GaugeAnnotationBeanPostProcessor.java
@@ -1,20 +1,22 @@
 package com.yammer.metrics.spring;
 
 import com.yammer.metrics.annotation.Gauge;
+import com.yammer.metrics.core.MetricName;
 import com.yammer.metrics.core.MetricsRegistry;
 import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.core.Ordered;
 import org.springframework.util.ReflectionUtils;
+import org.springframework.util.ReflectionUtils.FieldCallback;
 import org.springframework.util.ReflectionUtils.MethodCallback;
-import org.springframework.util.ReflectionUtils.MethodFilter;
 
+import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 
 public class GaugeAnnotationBeanPostProcessor implements BeanPostProcessor, Ordered {
 
-    private static final MethodFilter filter = new AnnotationMethodFilter(Gauge.class);
+    private static final AnnotationFilter filter = new AnnotationFilter(Gauge.class);
 
     private final MetricsRegistry metrics;
     private final String scope;
@@ -31,16 +33,30 @@ public class GaugeAnnotationBeanPostProcessor implements BeanPostProcessor, Orde
 
     @Override
     public Object postProcessAfterInitialization(final Object bean, String beanName) throws BeansException {
-        ReflectionUtils.doWithMethods(AopUtils.getTargetClass(bean), new MethodCallback() {
+        final Class<?> targetClass = AopUtils.getTargetClass(bean);
+
+        ReflectionUtils.doWithFields(targetClass, new FieldCallback() {
+            @Override
+            public void doWith(Field field) throws IllegalArgumentException, IllegalAccessException {
+                final Gauge gauge = field.getAnnotation(Gauge.class);
+                final String name = gauge.name().isEmpty() ? field.getName() : gauge.name();
+                metrics.newGauge(field.getDeclaringClass(),
+                                 name,
+                                 new GaugeField(bean, field));
+            }
+        }, filter);
+
+        ReflectionUtils.doWithMethods(targetClass, new MethodCallback() {
             @Override
             public void doWith(Method method) throws IllegalArgumentException, IllegalAccessException {
                 if (method.getParameterTypes().length == 0) {
                     final Gauge gauge = method.getAnnotation(Gauge.class);
                     final String name = gauge.name().isEmpty() ? method.getName() : gauge.name();
-                    metrics.newGauge(method.getDeclaringClass(),
-                                     name,
-                                     scope,
-                                     new GaugeMethod(bean, method));
+                    final String group = MetricName.chooseGroup(gauge.group(), method.getDeclaringClass());
+                    final String type = MetricName.chooseType(gauge.type(), method.getDeclaringClass());
+                    final MetricName metricName = new MetricName(group, type, name, scope);
+
+                    metrics.newGauge(metricName, new GaugeMethod(bean, method));
                 } else {
                     throw new IllegalStateException("Method " + method.getName() + " is annotated with @Gauge but requires parameters.");
                 }