+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-v2.0.0: TBD
+v2.0.2: Feb 09 2012
+===================
+
+* `InstrumentationModule` in `metrics-guice` now uses the default `MetricsRegistry` and
+  `HealthCheckRegistry`.
+
+
+v2.0.1: Feb 08 2012
+===================
+
+* Fixed a concurrency bug in `JmxReporter`.
+
+
+v2.0.0: Feb 07 2012
 ===================
 
 * Upgraded to Jackson 1.9.4.
@@ -7,6 +20,10 @@ v2.0.0: TBD
 * Fixed the inability to start and stop multiple reporter instances.
 * Switched to using a backported version of `ThreadLocalRandom` for `UniformSample` and
   `ExponentiallyDecayingSample` to reduce lock contention on random number generation.
+* Removed `Ordered` from `TimedAnnotationBeanPostProcessor` in `metrics-spring`.
+* Upgraded to JDBI 2.31.1.
+* Upgraded to Ehcache 2.5.1.
+* Added `#timerContext()` to Scala `Timer`.
 
 
 v2.0.0-RC0: Jan 19 2012